+++ CHANGES.rst
@@ -8,6 +8,8 @@ in progress
 - Move all services into module namespace
 - Fix PeriodicThread
 - Add "requests" module as a core requirement to "setup.py" as it is a common module used by many services
+- Add commands "mqttwarn make-config" and "mqttwarn make-samplefuncs"
+  for generating a mqttwarn.ini or a samplefuncs.py, respectively.
 
 
 .. _mqttwarn-0.7.0: